feat(tokens): add hovered color to format
