+++ packages/bezier-tokens/scripts/lib/transform.ts
@@ -1,8 +1,6 @@
 import type { Named, Transform } from 'style-dictionary'
-import tinycolor from 'tinycolor2'
 
-import { HOVERED } from './constants'
-import { clip, extractNumber, toCSSDimension } from './utils'
+import { extractNumber, toCSSDimension } from './utils'
 
 type CustomTransform = Named<Transform<unknown>>
 type Transforms = Record<string, CustomTransform>
@@ -105,63 +103,5 @@ export const CSSTransforms = {
       `linear-gradient(90deg, ${value
         .map(({ color, position }) => `${color} ${position}`)
         .join(', ')})`,
-  },
-  makeHoveredColor: {
-    name: `custom/css/${HOVERED}/functional-color`,
-    type: 'value',
-    transitive: true,
-    matcher: ({ type, filePath }) =>
-      type === 'color' && filePath.includes('functional'),
-    transformer: ({ value, filePath }) => {
-      function getHoveredColor(value: string, theme: 'dark' | 'light') {
-        const color = tinycolor(value)
-        const { h, s, l, a } = color.toHsl()
-
-        let alpha = a
-        let lightness = l
-        let saturation = s
-
-        if (a === 0) {
-          alpha = 0.1
-        } else if (a < 0.2) {
-          alpha = alpha * 1.5
-        }
-
-        if (theme === 'light') {
-          if (l <= 0.17) {
-            lightness = (l + 0.07) * 1.1
-            saturation += 0.05
-          } else {
-            lightness *= 0.93
-            saturation -= 0.03
-          }
-        } else {
-          if (l >= 0.83) {
-            lightness = (lightness - 0.2) * 0.98
-            saturation -= 0.03
-          } else {
-            lightness = (lightness + 0.04) * 1.005
-            saturation += 0.05
-          }
-        }
-
-        if (s <= 0.1 || s >= 0.9) {
-          saturation = s
-        }
-
-        const res = tinycolor.fromRatio({
-          h,
-          s: clip(saturation),
-          l: clip(lightness),
-          a: clip(alpha),
-        })
-
-        return res.toHex8String()
-      }
-
-      return filePath.includes('dark-theme')
-        ? getHoveredColor(value, 'dark')
-        : getHoveredColor(value, 'light')
-    },
   },
 } satisfies Transforms
