+++ packages/bezier-tokens/scripts/lib/utils.ts
@@ -1,3 +1,6 @@
+import { type TransformedToken } from 'style-dictionary'
+import tinycolor from 'tinycolor2'
+
 import { HOVERED } from './constants'
 
 export const toCamelCase = (str: string) =>
@@ -11,6 +14,63 @@ export const extractNumber = (str: string) =>
 export const toCSSDimension = (value: string) =>
   /^0[a-zA-Z]+$/.test(value) ? 0 : value
 
-export const clip = (value: number) => Math.min(Math.max(value, 0), 1)
+const clip = (value: number) => Math.min(Math.max(value, 0), 1)
 
 export const isHoveredTransformName = (name: string) => name.includes(HOVERED)
+
+const getHoveredColor = (value: string, theme: 'dark' | 'light') => {
+  const color = tinycolor(value)
+  const { h, s, l, a } = color.toHsl()
+
+  let alpha = a
+  let lightness = l
+  let saturation = s
+
+  if (a === 0) {
+    alpha = 0.1
+  } else if (a < 0.2) {
+    alpha = alpha * 1.5
+  }
+
+  if (theme === 'light') {
+    if (l <= 0.17) {
+      lightness = (l + 0.07) * 1.1
+      saturation += 0.05
+    } else {
+      lightness *= 0.93
+      saturation -= 0.03
+    }
+  } else {
+    if (l >= 0.83) {
+      lightness = (lightness - 0.2) * 0.98
+      saturation -= 0.03
+    } else {
+      lightness = (lightness + 0.04) * 1.005
+      saturation += 0.05
+    }
+  }
+
+  if (s <= 0.1 || s >= 0.9) {
+    saturation = s
+  }
+
+  const res = tinycolor.fromRatio({
+    h,
+    s: clip(saturation),
+    l: clip(lightness),
+    a: clip(alpha),
+  })
+
+  return res.toHex8String()
+}
+
+export const getHoveredColorToken = (
+  token: TransformedToken
+): TransformedToken => {
+  const theme = token.filePath.includes('darkTheme') ? 'dark' : 'light'
+  return {
+    ...token,
+    name: `${token.name}-hovered`,
+    value: getHoveredColor(token.value, theme),
+  }
+}
